+++ apps/catalog/catalog/src/app/marketplace/marketplace.module.ts
@@ -41,10 +41,6 @@ const routes: Routes = [{
       loadChildren: () => import('./movie/search/search.module').then(m => m.MarketplaceSearchModule)
     },
     {
-      path: 'search-new',
-      loadChildren: () => import('./movie/search-new/search.module').then(m => m.MarketplaceSearchModule)
-    },
-    {
       path: 'activity',   // List of notifications
       loadChildren: () => import('../activity/activity.module').then(m => m.ActivityModule)
     },
