--- apps/catalog/catalog/src/app/marketplace/movie/search-new/search.component.html
@@ -1,127 +1,0 @@
-<title-list [titles$]="movieSearchResults$">
-
-  <!-- SORT -->
-  <ng-template titleSort>
-
-    <mat-form-field class="sort">
-      <mat-label>Sort by</mat-label>
-      <mat-select [formControl]="sortByControl">
-        <mat-option *ngFor="let option of sortOptions" [value]="option">
-          {{option}}
-        </mat-option>
-      </mat-select>
-    </mat-form-field>
-
-  </ng-template>
-
-
-  <!-- SEARCH BAR -->
-  <ng-template titleSearch>
-
-    <form [formGroup]="filterForm">
-      <mat-form-field appearance="outline">
-        <mat-label>Search for titles, directors, keywords, ...</mat-label>
-        <mat-icon matPrefix svgIcon="magnifying_glasses"></mat-icon>
-        <input matInput type="text" formControlName="query">
-      </mat-form-field>
-    </form>
-
-  </ng-template>
-
-
-  <!-- FILTERS (ng-content) -->
-  <title-filter>
-
-    <ng-template filter label="TITLE METADATA" section></ng-template>
-
-    <ng-template filter label="Library / Line-up">
-      <static-check-boxes fxLayout="row" fxLayoutAlign="space-between" [form]="filterForm.storeType" model="STORE_TYPE"></static-check-boxes>
-    </ng-template>
-
-    <ng-template filter label="Genres">
-      <chips-autocomplete [form]="filterForm.genres" model="GENRES"></chips-autocomplete>
-    </ng-template>
-
-    <ng-template filter label="Country of Origin">
-      <chips-autocomplete [form]="filterForm.originCountries" placeholder="Country of Origin" model="TERRITORIES"></chips-autocomplete>
-    </ng-template>
-
-    <ng-template filter label="Language & Version">
-      <title-language-filter [languagesFilterForm]="filterForm.languages"></title-language-filter>
-    </ng-template>
-
-    <ng-template filter label="Production Status">
-      <static-check-boxes fxLayout="column" [form]="filterForm.productionStatus" model="MOVIE_STATUS"></static-check-boxes>
-    </ng-template>
-
-    <ng-template filter label="Minimal Budget">
-      <title-estimated-budget-filter fxLayout="column" [form]="filterForm.minBudget"></title-estimated-budget-filter>
-    </ng-template>
-
-    <ng-template filter label="AVAILS" section>
-      <p class="mat-caption">You need to fill in Terms, Territory and Media before applying the Avails filter</p>
-    </ng-template>
-
-    <!-- TODO -->
-    <ng-template filter label="Term">
-      <p>TODO</p>
-    </ng-template>
-
-    <!-- TODO -->
-    <ng-template filter label="Territory">
-      <p>TODO</p>
-    </ng-template>
-
-    <!-- TODO -->
-    <ng-template filter label="Media">
-      <p>TODO</p>
-    </ng-template>
-
-    <!-- TODO -->
-    <ng-template filter label="Exclusivity">
-      <p>TODO</p>
-    </ng-template>
-
-    <!-- TODO -->
-    <ng-template filter label="" section>
-      <button mat-stroked-button color="primary">Apply Avails Filter</button>
-    </ng-template>
-
-  </title-filter>
-
-
-  <!-- TITLE CARD -->
-  <ng-template titleCard let-title>
-
-    <movie-card [movie]="title" [link]="['/c/o/marketplace/title',  title.id]">
-      <!-- Information -->
-      <ng-container movieInformation>
-        <mat-card-subtitle>
-          <span class="mat-caption">
-            {{ title.main.directors | displayName }}
-          </span>
-        </mat-card-subtitle>
-      </ng-container>
-
-      <!-- Action -->
-      <ng-container movieCardAction>
-        <div fxLayout="row" fxLayoutAlign="end">
-          <wishlist-button [movieId]="title.id" small></wishlist-button>
-        </div>
-      </ng-container>
-    </movie-card>
-
-  </ng-template>
-
-
-  <!-- LIST ITEM CARD -->
-  <ng-template titleListItem let-title>
-
-    <!-- WRITE HERE WHAT YOU WANT TO INJECT-->
-    <!-- TODO use the real movie list item if it exists-->
-    <p>{{title.main.title.original}}</p>
-
-  </ng-template>
-
-
-</title-list>
