--- apps/catalog/catalog/src/app/marketplace/movie/search-new/search.component.scss
@@ -1,35 +1,0 @@
-@import '~@angular/flex-layout/mq';
-
-mat-form-field {
-  width: 100%;
-}
-
-title-filter {
-  static-check-boxes, title-estimated-budget-filter {
-    padding: 0 16px;
-  }
-}
-
-movie-card {
-  max-width: 25%;
-  padding-bottom: 24px;
-}
-
-@include layout-bp(xs) {
-  movie-card {
-    max-width: 65%;
-  }
-}
-
-@include layout-bp(sm) {
-  movie-card {
-    max-width: 45%;
-  }
-}
-
-@include layout-bp(md) {
-  movie-card {
-    max-width: 35%;
-    padding-bottom: 16px;
-  }
-}
